+++ app/build.gradle
@@ -20,7 +20,7 @@ android {
 }
 
 ext {
-    supportLibVersion = '25.1.0'  // variable that can be referenced to keep support libs consistent
+    supportLibVersion = '25.3.1'  // variable that can be referenced to keep support libs consistent
 }
 
 dependencies {
