+++ app/src/main/java/com/kinesioapp/kinesioapp/FrontPageFragment.java
@@ -1,10 +1,8 @@
 package com.kinesioapp.kinesioapp;
 
 import android.app.Fragment;
-import android.app.FragmentManager;
 import android.os.Bundle;
 import android.support.design.widget.TabLayout;
-import android.support.v13.app.FragmentPagerAdapter;
 import android.support.v4.view.ViewPager;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -25,8 +23,6 @@ public class FrontPageFragment extends Fragment {
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-
-
     }
 
     @Override
@@ -44,35 +40,10 @@ public class FrontPageFragment extends Fragment {
         ButterKnife.bind(this, view);
 
         // Get the ViewPager and set it's PagerAdapter so that it can display items
-        viewPager.setAdapter(new MyAdapter(getChildFragmentManager()));
+        viewPager.setAdapter(new TabViewPagerAdapter(getChildFragmentManager()));
 
         // Give the TabLayout the ViewPager
         tabLayout.setupWithViewPager(viewPager);
-    }
-
-    public static class MyAdapter extends FragmentPagerAdapter {
-
-        public MyAdapter(FragmentManager fm) {
-            super(fm);
-        }
-
-        @Override
-        public int getCount() {
-            return 4;
-        }
-
-        @Override
-        public Fragment getItem(int position) {
-            Bundle args = new Bundle();
-            args.putInt(FavoritesFragment.POSITION_KEY, position);
-            return FavoritesFragment.newInstance(args);
-        }
-
-        @Override
-        public CharSequence getPageTitle(int position) {
-            return "Child Fragment " + position;
-        }
-
     }
 
 }
