+++ app/src/main/java/com/kinesioapp/kinesioapp/RecentFragment.java
@@ -1,9 +1,7 @@
 package com.kinesioapp.kinesioapp;
 
-import android.content.Context;
-import android.net.Uri;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
+import android.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -12,6 +10,12 @@ public class RecentFragment extends Fragment {
 
     public RecentFragment() {
         // Required empty public constructor
+    }
+
+    public static RecentFragment newInstance(Bundle args) {
+        RecentFragment fragment = new RecentFragment();
+        fragment.setArguments(args);
+        return fragment;
     }
 
     @Override
