+++ app/src/main/java/com/kinesioapp/kinesioapp/TabViewPagerAdapter.java
@@ -1,0 +1,37 @@
+package com.kinesioapp.kinesioapp;
+
+import android.app.Fragment;
+import android.app.FragmentManager;
+import android.os.Bundle;
+import android.support.v13.app.FragmentPagerAdapter;
+
+public class TabViewPagerAdapter extends FragmentPagerAdapter {
+
+    public TabViewPagerAdapter(FragmentManager fm) {
+        super(fm);
+    }
+
+    @Override
+    public int getCount() {
+        return 2;
+    }
+
+    @Override
+    public Fragment getItem(int position) {
+        Bundle args = new Bundle();
+        if (position == 0) {
+            return FavoritesFragment.newInstance(args);
+        } else {
+            return RecentFragment.newInstance(args);
+        }
+    }
+
+    @Override
+    public CharSequence getPageTitle(int position) {
+        if (position == 0) {
+            return "Favorites";
+        } else {
+            return "Recent";
+        }
+    }
+}
