+++ app/src/main/res/layout/fragment_favorites.xml
@@ -6,14 +6,6 @@
     android:layout_height="match_parent">
 
     <TextView
-        android:id="@+id/textViewPosition"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:gravity="center"
-        android:text="This is a test"
-        android:textSize="50sp" />
-
-    <TextView
         android:id="@+id/favoritesText"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
