+++ app/src/main/res/layout/fragment_front_page.xml
@@ -9,11 +9,15 @@
         android:id="@+id/tabLayout"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:background="?attr/colorPrimary"/>
+        app:tabTextColor="@color/gray"
+        app:tabSelectedTextColor="@color/white"
+        android:background="@color/colorPrimary"
+        />
 
     <android.support.v4.view.ViewPager
         android:id="@+id/viewPager"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"/>
+        android:layout_height="match_parent"
+        />
 
 </LinearLayout>
