+++ app/src/main/res/values/colors.xml
@@ -1,6 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#3F51B5</color>
-    <color name="colorPrimaryDark">#303F9F</color>
-    <color name="colorAccent">#FF4081</color>
+    <color name="colorPrimary">#02CA37</color>
+    <color name="colorPrimaryDark">#01A02C</color>
+    <color name="colorAccent">#E5E4D7</color>
+
+    <color name="white">#FFFFFF</color>
+    <color name="gray">#DBDBDB</color>
 </resources>
